Remove disabled check for api_endpoint_url
When remotely deploying the applictation the url in the api_endpoint_url field is set to e.g. http://merchant-sample-cheapest:5003 and the field is disabled. This url is incorrect as it should probably point to the url where the application is remote deployed but cannot be changed without manually removing the disabled flag from the input field with the web inspector.